requirements.txt
fastmcp
httpx
beautifulsoup4
python-dotenv
pydantic
selenium
webdriver-manager
lxml
